update game text struct
update game text and split struct and enum
